Protocol changes from http:// to git:// Removed lowercase conversion Added a file_exists() check for rm

+++ lib/GitRepository.class.php
@@ -44,7 +44,7 @@ class GitRepository
 	}
 	
 	public function update(){
-	  exec(escapeshellcmd('rm -rf ' . $this->getPath()));
+	  if (file_exists($this->getPath())) exec(escapeshellcmd('rm -rf ' . $this->getPath()));
 		return $this->checkout();
 	}
